@@ -13,17 +13,24 @@
 	"dependencies": {
 		"@babel/core": "^7.6.4",
 		"@babel/preset-env": "^7.6.3",
+		"@material-ui/core": "^4.5.1",
+		"axios": "^0.19.0",
 		"babel-loader": "^8.0.6",
+		"material-ui": "^0.20.2",
+		"bootstrap": "^4.3.1",
+		"css-loader": "^3.2.0",
 		"express": "^4.17.1",
 		"react": "^16.10.2",
+		"react-bootstrap": "^1.0.0-beta.14",
 		"react-dom": "^16.10.2",
+		"react-redux": "^7.1.1",
 		"react-star-ratings": "^2.3.0",
-		"webpack": "^4.41.2",
 		"redux": "^4.0.4",
 		"redux-devtools-extension": "^2.13.8",
 		"redux-promise": "^0.6.0",
 		"redux-thunk": "^2.3.0",
-		"axios": "^0.19.0"
+		"style-loader": "^1.0.0",
+		"webpack": "^4.41.2"
 	},
 	"devDependencies": {
 		"@babel/preset-env": "^7.6.3",
@@ -33,7 +40,7 @@
 		"express": "^4.17.1",
 		"jest": "^24.9.0",
 		"nodemon": "^1.19.4",
-		"react": "^16.10.2",
+		"react": "^16.11.0",
 		"react-dom": "^16.10.2",
 		"react-test-renderer": "^16.10.2",
 		"webpack-cli": "^3.3.9"
