@@ -3,10 +3,13 @@ import ReactDOM from "react-dom";
 import App from "./components/App.jsx";
 import store from "./store";
 import newMainProduct from "./actions/newMainProduct";
-// adding to enable async/await with babel
-// const babelRegister = require("babel-core/register");
-// const babelPolyfill = require("babel-polyfill");
+import { Provider } from "react-redux";
 
-store.dispatch(newMainProduct(1));
+store.dispatch(newMainProduct(3));
 
-ReactDOM.render(<App />, document.getElementById("app"));
+ReactDOM.render(
+	<Provider store={store}>
+		<App />
+	</Provider>,
+	document.getElementById("app"),
+);
