@@ -1,10 +1,13 @@
 import React, { Component } from "react";
+import Overview from "./Overview/Overview.jsx";
+import R_R from "./R_R/R_R.jsx";
 
 export class App extends Component {
 	render() {
 		return (
 			<div>
-				<h1>Hello World!</h1>
+				<Overview />
+				<R_R />
 			</div>
 		);
 	}
