@@ -9,7 +9,7 @@ const getAvgRating = (ratings) => {
 		ratings[4] * 4 +
 		ratings[5] * 5;
 	const totalVotes =
-		ratings[0] + ratings[1] + ratings[2] + ratings[3] + ratings[4] + ratings[5];
+		ratings[1] + ratings[2] + ratings[3] + ratings[4] + ratings[5];
 	return totalStars / totalVotes;
 };
 
@@ -19,7 +19,9 @@ const mainProduct = (state = initialState.mainProduct, action) => {
 			return Object.assign({}, state, {
 				...action.productObj,
 				styles: action.styles.results,
-				avgReview: getAvgRating(action.reviewMetaData.ratings),
+				avg_review: getAvgRating(action.reviewMetaData.ratings),
+				reviews: action.reviews,
+				reviewMetaData: action.reviewMetaData
 			});
 		default:
 			return state;
