@@ -1,10 +1,10 @@
 import Redux from "redux";
 import initialState from "./initialState";
 
-const relatedProducts = (state = [initialState.relatedProduct], action) => {
+const relatedProducts = (state = [initialState.emptyProduct], action) => {
 	switch (action.type) {
-		case "":
-			return;
+		case "UPDATE_RELATED_PRODUCTS":
+			return action.relatedProducts;
 		default:
 			return state;
 	}
